Better logging

+++ source/commands/tickets.ts
@@ -1,13 +1,15 @@
-import { CategoryChannel, DMChannel, Guild, Message, NewsChannel, TextChannel } from 'discord.js';
-import { executeCommand, argsToString, generateValidationMessage, getCommand } from '../util/parsing';
+import { CategoryChannel, DMChannel, Guild, Message, MessageAttachment, NewsChannel, TextChannel } from 'discord.js';
+import { executeCommand, argsToString, generateValidationMessage, getCommand, stringToName } from '../util/parsing';
 import { Command } from '../models/command';
 
 const TICKET_CATEGORY_NAME = 'Tickets';
+const TICKET_LOG_NAME = 'ticket-logs';
+const TICKET_LOG_TOPIC_TEXT = 'Logs of every ticket closed';
 const TICKET_TOPIC_TEXT = 'To close the ticket type "!ticket close OPTIONAL_REASON". To add another user type "!ticket add NAME".';
 
 async function getTicketsCategory(guild: Guild): Promise<CategoryChannel> {
     let ticketsCategory = await guild.channels.cache.find(channel =>
-        channel.type == 'category' && channel.name == TICKET_CATEGORY_NAME
+        channel.type === 'category' && channel.name === TICKET_CATEGORY_NAME
     ) as CategoryChannel;
     if (!ticketsCategory) {
         ticketsCategory = await guild.channels.create(
@@ -41,14 +43,20 @@ async function createTicketTextChannel(ticketName: string, category: CategoryCha
 }
 
 export async function createTicket(command: Command, args: string[], message: Message): Promise<void> {
-    const ticketName = args[0];
+    let ticketName = '';
+    if (args[0]) {
+        ticketName = args[0];
+    } else {
+        ticketName = message.author.username + '-unnamed';
+    }
     const guild = message.guild as Guild;
-    if (ticketName && ticketName.length && ticketName.length <= 100) {
+    if (ticketName && ticketName.length && ticketName.length <= 100 && stringToName(ticketName) !== TICKET_LOG_NAME) {
         const category = await getTicketsCategory(guild);
         const ticketChannel = await createTicketTextChannel(ticketName, category, message, guild);
 
-        await ticketChannel.send(`Ticket "${ticketName}" opened by ${message.author.tag} on ${message.createdAt}`);
-        await message.delete();
+        await ticketChannel.send(`Ticket "${ticketName}" opened by ${message.author.tag}`);
+        await message.delete(); // TODO: delay this (delete after 15/30 sec)
+        // TODO: Link them to the channel, better starting message in the ticket with a mention and an embebbed command section
     } else {
         await message.channel.send(generateValidationMessage(command));
     }
@@ -56,23 +64,65 @@ export async function createTicket(command: Command, args: string[], message: Me
 
 async function isTicketChannel(channel: TextChannel | DMChannel | NewsChannel): Promise<boolean> {
     return channel.type === 'text' &&
-        channel.topic == TICKET_TOPIC_TEXT &&
+        channel.topic === TICKET_TOPIC_TEXT &&
         channel.parent !== null &&
         channel.parent.name === TICKET_CATEGORY_NAME;
 }
 
+async function getTicketLogChannel(category: CategoryChannel, guild: Guild): Promise<TextChannel> {
+    let ticketLogChannel = await guild.channels.cache.find(channel =>
+        channel.parent === category && channel.type === 'text' && channel.name === TICKET_LOG_NAME
+    ) as TextChannel;
+    if (!ticketLogChannel) {
+        ticketLogChannel = await guild.channels.create(
+            TICKET_LOG_NAME,
+            {
+                type: 'text',
+                topic: TICKET_LOG_TOPIC_TEXT,
+                position: 0,
+                parent: category,
+            }
+        );
+        await ticketLogChannel.lockPermissions(); // Sync with category permissions
+    }
+    return ticketLogChannel;
+}
+
+async function channelToText(channel: TextChannel | DMChannel | NewsChannel): Promise<Buffer> {
+    const messages = await channel.messages.fetch();
+    const messageArray: string[] = [];
+    messages.forEach(message => {
+        messageArray.push(`[${message.createdAt}] ${message.author.username}: ${message.content}`);
+    });
+    return Buffer.from(messageArray.reverse().join('\n'), 'utf-8');
+}
+
 export async function closeTicket(command: Command, args: string[], message: Message): Promise<void> {
     if (await isTicketChannel(message.channel)) {
-        const reason = args[0];
-        if (reason) {
-            await message.channel.send(`Ticket closed by ${message.author.tag} on ${message.createdAt} for reason: "${reason}"`);
+        const ticketChannel = (message.channel as TextChannel);
+        const guild = message.guild as Guild;
+        const category = await getTicketsCategory(guild);
+        const ticketLogChannel = await getTicketLogChannel(category, guild);
+        const textLog = await channelToText(ticketChannel);
+
+        const attachment = new MessageAttachment(textLog, `${ticketChannel.name}-log.txt`);
+
+        const uniqueTicketUsers = ticketChannel.members.filter(member => !ticketLogChannel.members.has(member.id));
+        const uniqueTicketUsersNames: string[] = [];
+        uniqueTicketUsers.forEach(user => {
+            uniqueTicketUsersNames.push(user.displayName + '\'s');
+        });
+
+        let closeMessage = `${uniqueTicketUsersNames.join(' & ')} ${!uniqueTicketUsers.size ? 'T' : 't'}icket closed by ${message.author.tag}`;
+        if (args[0]) {
+            closeMessage += ` for reason: "${args[0]}"`;
         }
-        // Create archive text channel (use getTicketCategory())
-        // TODO: Log ticket as .txt file
-        // send messages with the .txt file to the user who created/were added (this will require a database)
-        // look up how to do asyncronous message sending
-        // Post message in the channel
-        await message.channel.delete();
+
+        await ticketLogChannel.send(closeMessage, attachment);
+        uniqueTicketUsers.forEach(user => {
+            user.send(closeMessage, attachment);
+        });
+        await ticketChannel.delete();
     } else {
         await message.channel.send('Can only close Ticket channels');
     }
